backend/api/requirements.txt
flask==2.*
google-cloud-firestore==2.*
google-cloud-storage==2.*
gunicorn==21.*